Docker/requirements.txt
pillow
SimpleITK
imagecodecs
numpy
torchvision==0.15.1
fastai==2.7.12
timm==0.9.2